+++ app/(free-board)/_components/article-modal/schema/index.ts
@@ -1,6 +1,7 @@
 import { ObjectSchema, mixed, object, string } from "yup";
 
 import { FormType } from "..";
+import { EditFormType } from "../edit-article-modal";
 
 function checkIfFilesAreTooBig(file: File): boolean {
   let valid = true;
@@ -19,3 +20,11 @@ export const articleFormSchema: ObjectSchema<FormType> = object().shape({
     .required("이미지를 꼭 업로드해주세요.")
     .test("fileSize", "파일 사이즈가 너무 큽니다", checkIfFilesAreTooBig),
 });
+
+export const editArticleFormSchema: ObjectSchema<EditFormType> = object().shape(
+  {
+    title: string().required("제목을 꼭 입력해주세요!"),
+    content: string().required("내용을 꼭 입력해주세요!"),
+    image: string().required("이미지를 꼭 업로드해주세요."),
+  },
+);
